requirements.txt
azure-functions
azure-functions-durable
azure-storage-blob
opencv-python